+++ src/i18n.ts
@@ -9,6 +9,16 @@ interface RequiredLocales
 // プロジェクトごとの文字列リソースはこの型を使って定義のこと
 export type LocalizedMessages = Record<string, RequiredLocales>;
 
+
+export const COMMON_TEXTS: LocalizedMessages =
+{
+	showErrorDetailButtonCaption:
+	{
+		ja: '詳細を表示',
+		en: 'Show Detail',
+	}
+};
+
 // 文字列リソースの特定のキーが見つからない場合のエラーメッセージ
 const KEY_NOT_FOUND: RequiredLocales =
 {