+++ src/i18nTexts.ts
@@ -61,7 +61,262 @@ const messages: LocalizedMessages =
 	{
 		ja: '(名無しグループ)',
 		en: '(Nameless Group)'
-	}
+	},
+	updateLigatureFailed:
+	{
+		en: 'Failed to write editor.fontLigatures into the user settings.',
+		ja: 'ユーザー設定への editor.fontLigatures の書き込みに失敗しました。'
+	},
+	ligatures:
+	{
+		ja: '合字',
+		en: 'Ligatures',
+	},
+	letters:
+	{
+		ja: '文字',
+		en: 'Letters',
+	},
+	numbers:
+	{
+		ja: '数値',
+		en: 'Numbers',
+	},
+	stylisticSets:
+	{
+		ja: 'デザインセット',
+		en: 'Stylistic Sets',
+	},
+	widthVariants:
+	{
+		ja: '幅の異なる字形',
+		en: 'Width Variants',
+	},
+	culturalVariants:
+	{
+		ja: '文化的に異なる字形',
+		en: 'Cultrual Variants',
+	},
+	verticalFeatures:
+	{
+		ja: '縦書き機能',
+		en: 'Vertical Features',
+	},
+	others:
+	{
+		ja: 'その他',
+		en: 'Others',
+	},
+	liga:
+	{
+		ja: '標準合字',
+		en: 'Standard Ligatures',
+	},
+	calt:
+	{
+		ja: '前後関係に依存する字形',
+		en: 'Contextual alternates',
+	},
+	dlig:
+	{
+		ja: '任意の合字',
+		en: 'Discretionary Ligatures',
+	},
+	smcp:
+	{
+		ja: 'スモールキャップス',
+		en: 'Small Caps',
+	},
+	c2sc:
+	{
+		ja: '大文字のスモールキャップス',
+		en: 'Capitals to Small Caps',
+	},
+	swsh:
+	{
+		ja: 'スワッシュ字形',
+		en: 'Swashes',
+	},
+	salt:
+	{
+		ja: 'デザインのバリエーション',
+		en: 'Stylistic alternates',
+	},
+	lnum:
+	{
+		ja: 'ライニング数字',
+		en: 'Lining Figures',
+	},
+	onum:
+	{
+		ja: 'オールドスタイル数字',
+		en: 'Oldstyle Figures',
+	},
+	pnum:
+	{
+		ja: 'プロポーショナル数字',
+		en: 'Proportional Figures',
+	},
+	tnum:
+	{
+		ja: '等幅数字',
+		en: 'Tabular Figures',
+	},
+	frac:
+	{
+		ja: '分数',
+		en: 'Fractions',
+	},
+	ordn:
+	{
+		ja: '上付き序数表記',
+		en: 'Ordinals',
+	},
+	ss:
+	{
+		ja: 'デザインのセット',
+		en: 'Stylistic Sets',
+	},
+	pwid:
+	{
+		ja: 'プロポーショナル字形',
+		en: 'Proportional Widths',
+	},
+	palt:
+	{
+		ja: 'プロポーショナルメトリクス',
+		en: 'Proportional Alternate Widths',
+	},
+	pkna:
+	{
+		ja: 'プロポーショナルかな',
+		en: 'Proportional Kana',
+	},
+	fwid:
+	{
+		ja: '等幅全角字形',
+		en: 'Full Widths',
+	},
+	hwid:
+	{
+		ja: '等幅半角字形',
+		en: 'Half Widths',
+	},
+	halt:
+	{
+		ja: '字幅半角メトリクス',
+		en: 'Alternate Half Widths',
+	},
+	twid:
+	{
+		ja: '等幅三分字形',
+		en: 'Third Widths',
+	},
+	qwid:
+	{
+		ja: '等幅四分字形',
+		en: 'Quarter Widths',
+	},
+	jp78:
+	{
+		ja: 'JIS78 字形',
+		en: 'JIS78 Forms',
+	},
+	jp83:
+	{
+		ja: 'JIS83 字形',
+		en: 'JIS83 Forms',
+	},
+	jp90:
+	{
+		ja: 'JIS90 字形',
+		en: 'JIS90 Forms',
+	},
+	jp04:
+	{
+		ja: 'JIS2004 字形',
+		en: 'JIS2004 Forms',
+	},
+	trad:
+	{
+		ja: '旧字体',
+		en: 'Traditional Forms',
+	},
+	ruby:
+	{
+		ja: 'ルビ用字形',
+		en: 'Ruby Notation Forms',
+	},
+	hkna:
+	{
+		ja: '横組み用かな',
+		en: 'Horizontal Kana Alternates',
+	},
+	nlck:
+	{
+		ja: '印刷標準字体',
+		en: 'NLC Kanji Forms',
+	},
+	nalt:
+	{
+		ja: '修飾字形',
+		en: 'Alternate Annotation Forms',
+	},
+	ital:
+	{
+		ja: 'イタリック',
+		en: 'Italics',
+	},
+	vkrn:
+	{
+		ja: '縦組みペアカーニング',
+		en: 'Vertical Kerning',
+	},
+	vert:
+	{
+		ja: '縦組み用字形',
+		en: 'Vertical Alternates',
+	},
+	vpal:
+	{
+		ja: '縦組みプロポーショナルメトリクス',
+		en: 'Proportional Alternate Vertical Metrics',
+	},
+	vhal:
+	{
+		ja: '縦組み字幅半角メトリクス',
+		en: 'Alternate Vertical Half Metrics',
+	},
+	vkna:
+	{
+		ja: '縦組み用かな',
+		en: 'Vertical Kana Alternates',
+	},
+	kern:
+	{
+		ja: 'カーニング',
+		en: 'Kerning',
+	},
+	ccmp:
+	{
+		ja: '字体組版／分解',
+		en: 'Glyph Composition/Decomposition',
+	},
+	locl:
+	{
+		ja: 'ローカライズの字形',
+		en: 'Localized Forms',
+	},
+	sups:
+	{
+		ja: '上付き文字',
+		en: 'Superscript',
+	},
+	subs:
+	{
+		ja: '下付き文字',
+		en: 'Subscript',
+	},
 };
 
 export default messages;